+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+
+## 1.3.0 - 2022-04-11
 ### Added
 - `e d` to edit entry description
 - `e t` to edit entry task